Refactor folder structure for `parsing` to `scheduling` in `tokenRemove.py`, `tokenModify.py`, and `tokenAdd.py` for improved modularity and organization.
