+++ parsing/tokenize.py
@@ -1,6 +1,4 @@
 import re
-
-from parsing.tokenAdd import TokenAdd
 from utils.timeUtils import toUnixTime, toSeconds
 
 """
